fix: move tests to functional folder
