+++ tests/functional/register.spec.ts
@@ -1,8 +1,8 @@
-import { randomUserData } from '../src/factories/user.factory';
-import { RegisterUser } from '../src/models/user.model';
-import { LoginPage } from '../src/pages/login.page';
-import { RegisterPage } from '../src/pages/register.page';
-import { WelcomePage } from '../src/pages/welcome.page';
+import { randomUserData } from '../../src/factories/user.factory';
+import { RegisterUser } from '../../src/models/user.model';
+import { LoginPage } from '../../src/pages/login.page';
+import { RegisterPage } from '../../src/pages/register.page';
+import { WelcomePage } from '../../src/pages/welcome.page';
 import { expect, test } from '@playwright/test';
 
 test.describe('Verify registration', () => {
